Add qa/967 to test the pmdapapi auto_enable functionality
In this testcase we take two dbpmda runs;

First, we run a check that the existing enable/disable functionality
works normally when papi.control.auto_enable is disabled (set to 0)

Second, we check that the auto_enable timeout works by setting it to
a low timeout (5 seconds), fetching a value, checking that
papi.control.status reports the correct timeout and a valid papi
VALUE.  After which we wait just over 5 seconds and re-fetch the
papi.control.status value, which (should) show the metric is no
longer being counted and that papi has stopped.

This second test implicitly checks the functionality of dynamically
starting a counter.