buildrules: need rm -f before ln -s
In the build this was hidden because the alternate named python and
perl files did not exist.

But the same rule is used in qa/src and there the targets may exist
and it was only a matter of time before this bit somone ... like me
a couple of days ago.