qa/115: handle missing controls for init scripts
When we don't have the moral equivalent of chkconfig, we can't
really disable a service, so running the init.d script directly
as this test does will always start the service.

Add some convoluted logic in the test to detect this case and
behave accordingly.

Was failing on i486 Gentoo 2.0.3 (vm05).