Alter qa/903 awk statement to account for lower possible metric counts
The number of available papi metrics varies based on the system being
run on.  Previously there would be a pmid for each possible metric, so
we could set the awk regex much higher.  At this point, limit it to 7
or greater, (one for each papi.control and one papi.available).