rc-proc.sh: fix minor systemctl-related regression
On some platforms, systemctl does not like our services and
redirects queries about their state to chkconfig, as in ...

  $ systemctl is-enabled pmlogger.service
  pmlogger.service is not a native service, redirecting to /sbin/chkconfig.
  Executing /sbin/chkconfig pmlogger --level=5
  pmlogger  on
  enabled

This is benign, _except_ that it confuses the is_chkconfig_on()
method and it does not return the true state of the service,
i.e. 0 for on or enabled, else 1 for off or disabled).

Fix this and add some debugging (under the control of
$VERBOSE_CONFIG when set to true) to report which of the twisty
passages is_chkconfig_on() is using to arrive at an answer.