pmda proc: rework existing per-cgroup metrics, and add new ones
Fairly large rewrite of the Linux cgroup metrics, moving from a
metric name encoding of cgroups to using instances (solving the
long-standing naming problem, with cgroups named such that they
could not be represented).

While there, I've audited the kernel to check the semantics for
many of these metrics, and updated several that were incorrect.
Added per-cgroup totals for the blkio statistics.

The code behind the proc.runq metrics has been tweaked a little
so that only a single sweep through /proc is made for a pmFetch
request with multiple proc metrics.

Finally, the code is much cleaner and easier to follow now, so
hopefully encourages addition of new cgroup metrics as they come
along.