libpcp_pmda: change e_ext_t to expose pmdaInterface
The e_ext_t struct is internal to libpcp_pmda, but it is exposed
via the e_ext field of the pmdaExt struct that is accessible to
every callback that a PMDA developer might choose to override.

This change replaces the pmda_interface field of e_ext_t with
a back pointer to the PMDA's controlling pmdaInterface struct
(from which pmda_interface can be found, but more importantly a
while lot more information is available).

Motivated by Frank's observation that the existing libpcp_pmda
does not allow a "fetch" callback to be independent of the
metrictab, which restricts some dynamic metric use cases where
the pmDesc is not in a metrictab, but is instantiatd by the PMDA's
"desc" callback.