papi pmda: use dynamic metric description
Using kenj's commit #f856e2c171, it becomes possible to avoid
allocating a metrictab[] in the pmda, when it is so easy to generate
each pmDesc on demand.  This is done by having the pmdaFetch()
function in libpcp_pmda fall back to the pmdaInterface->desc callback
(if set), if a metrictab[] was not specified during pmdaInit().

The papi pmda is converted to this scheme.  A few metric pmUnts are
corrected.