PMAPI extension: pmParseUnitsStr
We introduce a function that reverses pmUnitsStr (and then some), so
that users can specify desired metric unit/scale conversions in
somewhat free-form textual form.  The API function looks like this:

     int pmParseUnitsStr(const char *, pmUnits *, double *)

where the latter two arguments are the results.  The first output is
what one would expect.  The last parameter is a final arithmetic
multiplier, which may be simply 1.0 for a straight pmUnitsStr() format
text, or is the appropriate factor to compensate for a user asking for
conversion of a quantity to

     "10 megabytes / 15 minutes"