rc_pmcd: execute _pmda_setup in the background
Amer reports that use of the .NeedInstall mechanism for
PMDAs introduces longer image startup times due to the
rc_pmcd script performing PMDA installation serially.
There's no reason for that - this processing can be done
in the background as soon as pmcd has started (just like
we do with pmloggers in the rc_pmlogger script already).

Test qa/300 is tweaked to give a little more time before
verifying the .NeedInstall processing has / has not been
done.