pmda unbound: add initial tests for Unbound DNS resolver PMDA
This adds two tests for the new Unbound DNS resolver PMDA - one to
exercise the metric values being extracted from the unbound-control
command output, the other to exercise the Install/Remove process.
Found and fixed a bug in the latter, where an exception would be
raised if the executed command did not exist.

Also added a makefile for the unbound PMDA and generally stitched
it into the build/packaging, ready for the next release.