libpcp/pmcd: protocol-level support for containers
Generalisation of the existing authentication PDU into an
attribute PDU, and addition of support into pmcd allowing
clients to request that PMDAs provide information about a
named container.  At this stage clients can inform PMDAs,
via pmcd, that they are interested in a named container.

The pmcd and linux kernel PMDAs are updated to interface
version 6 here also, and attribute callbacks are added to
prepare them for responding to requests for metrics from
individual containers.