libpcp_pmda - add pmdaInterfaceMoved() function
The recent change to e_ext_t replaced the int pmda_interface field
(held a PMDA_INTERFACE_N value) with a back pointer to the base of
the pmdaInterface struct.

It was believed that this structure was only manipulated within
libpcp_pmda, but co-incident and non-deterministic failures of
qa/628 led to some suspicion that this assertion was not correct.

The e_ext_t struct IS only manipulated within libpcp_pmda, but the
the associated pmdaInterface struct is created outside libpcp_pmda,
and the back pointer ONLY works if the pmdaInterface struct is NOT
moved or relocated.

For most PMDAs this is not a problem.  But pmcd sometimes will shuffle
and array of pmdaInterface structs and sometimes realloc the whole array,
thereby both moving AND relocating the pmdaInterface struct.

The new pmdaInterfaceMoved() routine re-establishes the back pointer
after a pmdaInterface struct is moved.

Also updated the man pages for pmdaDaemon() and pmdaDSO() to
document the caveat that the pmdaInterface struct passed by pointer
to these routines is not expected to be moved, and if it is moved
pmdaInterfaceMoved() is your new best friend.