pmda root: a privileged pmcd co-process
Initial implementation of "pmdaroot", a helper co-process
for pmcd that runs with elevated privileges and is able to
(well, will be able to soon) offer services to pmcd and to
other PMDAs where root access is required.

See http://www.pcp.io/pipermail/pcp/2014-June/005030.html
for additional rationale.  I've implemented this as a PMDA
(rather than simply forking pmcd early and embedding all of
the code in pmcd itself, as considered earlier) as that has
a number of advantages:
- it can be switched off if there's a problem
- complete code and functional isolation from pmcd
- i.e. no changes required to pmcd to support it - safe!
- uses existing pmcd mechanisms to bootstrap, which means
  less code was needed
- ability to export metrics about its operation, such as
  the containers it has found and what state they're in.

I also recycled PMDA domain number 1 for this PMDA, because
IRIX is gone and this is hopefully going to become a widely
used piece of infrastructure over time, so it seems a good
candidate for using that number.  It is enabled-by-default
for new installations, Install/Remove scripts are provided
for now also (but maybe not in the long term).

This initial version provides a simple mechanism to discover
information about Docker containers, in preparation for the
next stages.  It will probably need to be enhanced to using
the Docker daemon at some point (via the remote Docker API &
AF_UNIX socket - see details on docker.io) instead of simply
grovelling around on-disk, but for now thats what it uses.

Support for exporting state about other styles of container
will also be added in due course I hope.  The pmdaroot code
has been written such that drop-in container "drivers" for
any container technology could be added (there are many on
Linux now, Solaris Zones, and so on).