pmdaroot/libpcp_pmda: interfaces for changing namespaces
Work-in-progress commit with the majority of the code behind
the PCP collector privilege escalation mechanism, and with a
Linux-kernel-namespace switching API for PMDAs with metrics
that need to be container namespace aware, as the first use
case.  The pmcd.hostname metric from the pmcd DSO PMDA has
been converted over to using the new API - others remaining
to be tackled include network.interface.*, filesys.*, ipc.*
and one or two others.  See also notes in the code as to some
other PMDAs that could use pmdaroot services (e.g. pmdaproc
restricting per-process indom to just processes from a given
containers cgroups, that sort of thing).

Committing it now, even though its not quite done, since it
has no impact on the code when not used (which it isn't yet)
and some people are interested in seeing how it fits together
while I'm away on vacation.  In the couple of places where
more coding will be needed "TODO" markers have been inserted
to guide the way.