Makepkgs: special case handling of VERSION.pcp
In the QA context, packages get respun often and for rpm in
particular this means the PACKAGE_BUILD needs to get incremented.

The qa/admin/pcp-daily script organizes this (using a modified
VERSION.pcp that is not committed to any git repo), but when
Makepkgs is using git-archive to create the source tarball the
modified VERSION.pcp file is not included.

Added some logic to handle this after the source tarball has been
unpacked.