pmdaproc.sh: sort pmdas based on domain and maintain space
Increase the chance that pmdaroot is the first PMDA started,
because soon other PMDAs will depend on it.  A nice simple
solution for this special case (IOW, its not a general issue
calling for a more general/invasive solution like a pmcd.conf
format extension).

While there, ensure the space between the access section that
is created during the build is not lost during Install/Remove
of PMDAs (cosmetic, harmless).