pmdaroot: pmdalinux namespace awareness, further QA work
Tweaks to the API, things beginning to work - clients connecting
and so on.  Converted the Linux kernel agent to be aware of user
requests for metric values tailored to individual containers (so
filesys.*, network.*, ipc.* etc now container aware).  Some small
tweaks to the PMDA API to pass in a file descriptor, instead of
using globals.  Allow pmdaroot to use a custom socket path.

Some QA coverage beginning to appear beyond the basics, and also
converted qa/024 over to simpler one-output variants.