pmlogextract: fix recent metadata regression
Recent pmlogextract changes to tighten metadata consistency checks
were triggering new failures in the daily log merging on bozo.

Eventually tracked down to a (yet to be resolved) pmlogger bug
where the archives contain two identical names for the same pmid.

Fix here is to expand the matchnames() function to handle cases
where the set of names for a metric may contain duplicates.