pmmgr: rpm packaging tweaks
Because pmmgr is configured by individual files, where their
non-existence can be a signal (such as to disable pmie).  This is a
problem if a later package (RPM) upgrade brings a new copy of
/etc/pcp/pmmgr, including the deliberately rm'd files.  To fix this,
rpm allows such config files to be marked %config(missingok,noreplace),
which is henceforth applied to the pmmgr ones.  While in the vicinity,
toned down the cautionary language used to describe the subrpm.