pmmgr fix: handle rapidly-changing ip addresses
It can happen that a pmmgr-monitored target changes its ip address so
quickly that it never appears 'down' between two poll intervals.  In
this case, pmmgr didn't treat the renumbered target as requiring
daemon stopping/restarting, and instead futilely tried to restarting
the daemons against the old address.

This is corrected by checking the new preferred-address for a host
against the previously preferred one, and treating a change as a
dead/alive transition.

Testing included multiple VMs on a network hand-changing addresses
via avahi and/or raw ip, to stress both libpcp discovery streams.
(Mechanising this test is unfortunately not obvious.)