pmParseUnitsStr: python binding
Add a pmapi context pmParseUnitsStr(string) -> (units,multiplier)
frontend to the python bindings.  Add test case, confirming a sample
success case, an error case, and a ctypes type-error case.  (The
latter is entertaining: if it were not for an isinstance() check in
the python binding, passing an integer parameter to the underlying
LIBPCP.pmParse$foo function could cast it to a char*, and cause a
SEGV.  Something worth checking on other python bindings.)