containers: rework pmdaroot PDUs, pmdalinux, pmdaproc updates
Change the division of labour that pmdaroot takes on.  Most of
the setns calls are moved into the Linux-specific PMDA, where
they really belong.  PDUs are added for looking up cgroups for
a given container, and pmdaproc has been updated to know about
containers and to use the cgroup PDU to restrict the instances
and values it returned.