qa/730 & 731: cgroup test rework
I am guessing the "cgroup files from a tar ball" testing exposes a
dependency on readdir() order and where tar whacks the files down
because I'm seeing lots of QA failures that relate to (a) the order
of the metrics from pminfo, and (b) the encoding of internal instance
ids for the various indoms here.

Filtering out (b) produces identical output, except the order is
all over the place.

Made changes to take control of the non-determinism ... tests passing
in lots of places now.