pmlogreduce: duplicate PMNS names allowed by default
Trickier change here as pmlogreduce does "interp" mode fetches from
the input archive and uses pmTraverseNameSpace() to enumerate the
names (and then PMIDs) for all the metrics in the input archive.

With possible duplicate names in the PMNS this was triggering a
horrible bug/design flaw in interp.c where if the same PMID appears
more than once in pmidlist[] passed to pmFetch, then interp.c can
get its linked lists in a royal knot.  This has not been a problem
in the past because pmOptFetch() and friends have pretty much
ensured than the same PMID does not appear more than once in
the pmidist[] passed to pmFetch().

Filter the metrics to ensure each PMID appears at most once in the
list of metrics to be fetched.

Use pmNameAll() to report all of the names of metric in the output
and to ensure all of the aliased names for each metric appear in the
metadata of the output archive.