qa/555: smack long standing failure
For most of this millenium (well it seems that way), qa/555 has
been failing on bozo.

Finally got to the bottom of it ... basically
	$ sudo -E service rsyslog <anycmd>
returns
<anycmd>: Unknown job: rsyslog
but
	$ sudo service rsyslog cmd
works.

We favour -E and $sudo is set to include -E in common.rc if the
sudo version supports this.

So we never actually restarted rsyslogd, which explains why the pmie
syslog messages were not found in /tmp/syslog.out!

Sigh, next .bad boy!