pmdaproc.sh & sample PMDA: last piece of duplicate names in PMNS work
With this commit, when a PMDA is "Install"ed there is an additional
check that the PMDA's PMNS (not the whole PMNS) does not contain any
duplicate names (or indeed other errors).

This puts back the original checks where a duplicate name in a PMDA's
PMNS was most likely to be a programmer error, rather than a deliberate
decision.

The sense of the check is controlled by a new $pmns_dupok variable
that defaults to false, but can be set
pmns_dupok=true
in a PMDA's Install script if the PMDA intentionally has duplicate
names in it's PMNS, e.g. the sample PMDA.