qa/837: was killing the primary pmlogger
Need to start pcp, not pmcd, and wait for pmcd and pmlogger.

Does not change this test, but does change state so that later
tests are not victims.  Specifically this test was causing qa/897
to fail in full qa runs, even though qa/897 is rock solid when
rechecked or run alone.