pmdaproc.sh: handle schizo PMDAs with daemon/Perl/Python implementations
Turns out "root" is only a good local PMNS if pmns.save does not
exist.  The Python and Perl install parts of pmdaproc.sh move
pmns aside (to pmns.save) and leave their own version in pmns.
In this case, we need to synthesize a "root" for the purposes of
doing the PMNS check.