src/pmdas/systemd/Install: need to be trickier about access controls
By default, this PMDA runs as the user "adm" ... this is sometimes
enough to get to the systemd logs.  If the user "adm" does not exist
we run the PMDA as root.

What this commit adds is a check when the user "adm" exists, but
cannot access the logs ... use journalctl to check this and run
PMDA as root if access is denied to the "adm" user.