pmatop: arrgh, got you! truncated output bug squashed
qa/722 has been failing on a few machines for a very long time.

I know nothing about Python, so I was avoiding it.  But I decided
to dig a little.

Turns out if you don't have /dev/mapper, then there is a unhandled
exception at the start of the LVM processing, and this forces
control out of the main loop iterating over metrics.  So the last
line of output you see is "PAG | ..".

It would probably be a good idea if someone who groks Python checks
my "fix" ... 8^)>