qa/common.check: make _cleanup_pmda() wait and check a bit more
After pmcd is restarted here, there is another possible restart
in the ./Install step ... make sure pmcd and pmlogger are running
before returning.

I am using the check.callback mechanism to check if the file
/var/lib/pcp/tmp/pmlogger/primary exists after each test is run
and qa/761 was tripping this check, sometimes.

This change removes the small window between the end of one test
and the start of the next where the primary pmlogger may not yet
have finished restarting.