pmdalinux: improve /proc/cpuinfo parser for irregular formats
For a long time the architectures with /proc/cpuinfo forms that
have global CPU information either before/after the individual
"processor:" lines have been problematic.  The last wrinkle was
on s390 where the global info precedes these lines (previously
we'd only seen archs with processor lines at the files head).

Rework things so that global information is saved on-stack in
the same way for all platforms (removing Alpha cpp conditional
code, and improving detection on Alpha at the same time - as
can be seen in qa/885).  The test 885 expected output has been
updated to include the new s390x output, which previously was
causing a sigsegv.