qa/040: tweak timing for pmlogger
With the changes to pmlogger to avoid the signal deadlock, it is
no longer possible to log reliably at a 5msec sample frequency.
(It never really was, but previously we never got out of the
interrupt handler!).  Logging at 20msec seems to work fine.

Fine-tune the number of logging samples and the checking iterations
to match.