python3: on-going work to support both python 2 and 3
Series of changes around python3 tooling, all should be noops
upstream for the existing python2 support though.

RPM packaging specfile work, requested by the python packagers
for fedora22 and rawhide.  Added a flag to allow python3 to be
set as default, and adjust dependencies and hash-bang lines as
appropriate.  If using a python3-only build root, we toss the
systemtap static probes support cos its python2-only for the
foreseeable future.

Add configure support for a PCP_PYTHON_PROG, letting us set the
preferred python/python3 binary to be set.  This is needed for
the python PMDA installs in particular.  We are now aggresively
going to default to using python3 if we're able to build for it
so we can start ironing out remaining issues (--without-python3
to ./configure if are faint of heart).

All python PMDAs updated with minor runtime updated for python3,
and all associated QA updated to be able to work both ways.  We
now use PCP_PYTHON_PROG to set the preferred QA python binary as
well.

QA 667 is fixed to not assume pcp2graphite is functional even in
just --help mode (cos its not with python3 syntax errors) - this
caused the test to _notrun when it should have failed.

Since python 2.6+ is nowadays mandated for PCP builds, some tests
have been simplified to remove checking for earlier versions.

Python3 is picky about whitespace (tabs vs spaces) - fixed many
places where this issue still lurked.  long() and int() are two
other common patterns of failure tackled here, as are numerous
string-encoding issues while interfacing to the C libraries.

All changes in the context of Red Hat bugzilla bug #1194324.