qa: rework detection of containers and setns syscall
Make the pmcd.feature.containers metric value depend on
whether the setns syscall is available, allowing finer-
grained detection in QA tests.

Rearrange the _check_containers qa/common helper to make
use of it.  Move the earlier fetch-from-every-PMDA check
to qa/359 as a third case of checking the world, but for
containers as well.