qa/752: harder to fix than killing a horde of zombies
Turns out that our gnulib-based code has evolved over time,
as has the official version that is used by date(1).

As a consequence, we may (on older platforms) have a fixed libpcp
but a (semantically) broken date(1) ... which causes the QA test
to fail.

This commit does a Larry Wall-like sniff test to determine if
date(1) is broken and then tweaks the filter to make the test
pass ... don't look unless you have a strong stomach.