python3: support pmns traversal callbacks with bytes->strings
One area where supporting native unicode python strings above
the python PMAPI wrapper was difficult is where a callback is
provided that is passed a string from C code.  pmTraversePMNS
is such a case, and we tackle it here via a little wrapper in
the C code to help.

Test qa/904 is added to exercise this case, and a couple of
the other namespace expansion/iteration cases in a more focused
manner than the general python API test provides.