rc: reinstate lost pmlogger verbosity for older init systems
Further steps toward pmie/pmlogger init script consistency, and
fix up some diagnostics around chkconfig state that pmie script
had but not pmlogger.  Tackles the Red Hat quirk with locations
of init scripts too.  None of this is useful for systemd but we
make life better on the other systems here - further work still
to be done in the systemd area (if thats possible, not clear).