pmwebd: add a "-i MIN-INTERVAL" option for graphite time-precision control
When pmlogger data is sampled more frequently than /60s, it is
desirable to let graphite render data at a higher time resolution.
Oversampling leads to wiggles due to interpolation FP artifacts, so is
undesirable.  The pmwebd invoker is best placed to judge the right
tradeoff, so is given an option to set the minimum time-step size.