libpcp memory corruption bug: __pmSendInstanceReq doing __pmFindPDU too small
An exact-size __pmFindPDU malloc implementation found this bug in
p_instance.c, wherein the `sizeof(need)' rather than `need' was
passed.  This corrupted memory behind the declared region, but the
problem was hidden because the clasical __pmFindPDU rounded up memory
allocations to 1K+, but large indom-name queries can trigger it.