pmcd: fix failure to recover from errors sending client attrs
Properly handle a failed attempt to send attributes to a PMDA
that has died - we need to mark this PMDA as such and ensure no
access continues for the client request that trips over it.

Adds test qa/956 which exercises the expected behaviour via the
daemon dynamic PMDA with a small addition (requesting attrs).