Add script to extend qa/926
rrdtool is very platform/arch specific. We provide a set of xml
files that can be converted to the appropriate format for new
architectures.

If you get the following error:

ERROR in RRDs::info .... This RRD was created on another architecture

Then : cd qa/ganglia && mk-gangliatest

Should generate new files for this architecture.