qa/common.check: teach _host_to_fqdn() to tolerate no domainname
There exist perfectly functional machines that have a non-fqdn
hostname(1), an empty domainname(1), and don't happen to show up in
DNS.  These may be fully reachable on the network via avahi or
getent/nsswitch/hosts, but several pcpqa tests (e.g., 651) reject
them.  But alas, even a _notrun would be better than what happens
currently:

This patch rejects the _host_to_fqdn() heuristic's attempt to suffix
the incoming hostname with `domainname`, if that result was "(none)".
Passing "(none)" back as part of a fqdn is a recipe for tears and qa
failure.