RHBZ 1199052: malloc/free misbehaviour in pmwebd when avahi client wasn't created.
The failed avahi service handle was freed after the error but was not
NULLed out and was also left on the list of active services. This
allowed the freed memory to be referenced again duing subsequent
avahi callbacks and also to be freed again when __pmAvahiUnadvertieService()
was called during shutdown.