pmdadm: Device Mapper (md) PMDA Initial Merge
Initial Device Mapper PMDA code which is a rewrite and merging of the
dm-thin and dm-cache PMDA in order to minimise PMDA proliferation.
Making use of stats given by dmsetup status for the cache, thin-pool
and thin targets we export the statistics for DM Cache, thin-pools
and thin provisioned volumes.

Providing metrics from the information given in their respected docs
cache.txt and thin-proivisioning.txt in the kernel docs/device-mapper
directory. There have been additional metrics added for the thin-pool
and thin-vol stats to give back the size of the pools/volumes. Have
also made changes as suggested by Nathan Scott to reduce the amount
of branching code in the fetch/instance functions.

QA is given by qa/657 which tests the PMDA without the need for local
thin-pools, volumes of caches. (Although the dmsetup tool is required
for the PMDA to install).

Have removed/renamed the non needed dm-thin and dm-cache PMDA's along
with their corresponding qa/tests and have used the PMID from dm-cache
as the PMID for the pmdadm.

Code checked with Coverity and has no reported issues.