pcp-verify: utility for checking collector installations
Adds new utility pcp-verify(1) which checks validity of a
collector installation, for various deployment scenarios.
Test qa/839 is added to check it.  The makefile generating
the archives from qa/archives is made to use wildcards now
rather than having to be updated for each new addition.  A
new qa group is added for all the pcp sub-commands also.