pcp-summary: refactor pcp(1) into its component entities
The role of the pcp(1) command is two-fold - its traditional
(summary) reporting role, and also as a sub-command frontend.
These are completely separate, and it turns out can be easily
made into separate entities - which this commit does via the
creation of a separate pcp-summary(1) sub-command.  This sub-
command is also the default, in the absence of any others.

There is essentially no functional change here, it is simply
a refactoring of code into two logically separate entities -
backward compatibility is preserved of course.