pmdadm: refactor code to remove _isQA completely
Also moved the instance-refresh code specific to each indom to
the source file associated with that area (dmthin/dmcache).

Removed struct wrappers for instances which seemed to be there
to add a copy of the instance name, which was set but not used
anywhere (reduces memory footprint by 4k for each instance) --
instance names are stored safely in the indom caches, so maybe
this was leftover from earlier work.

Removed previously unused metrictable_size function.