rc_pmcd: add .NeedRemove option for pmcd start script also
When transitioning metrics into a new PMDA from another, as
in the case of the dmcache and dm PMDAs, we need to be able
to switch an existing installation from one PMDA to another
in an automated fashion.

This commit extends the existing NeedInstall concept with a
NeedRemove equivalent, paving the way for dmcache installs
to be detected, ./Remove'd and then ./Install'd via pmdadm.