pmdumplog: new options and new log reading logic
New options -M (force mark records to be reported) and -xx (report
timestamps as offset in seconds from the start of the archive).

But the more substantive change is to replace calls to pmFetch()
and pmFetchArchive() with calls to __pmLogFetch() ... the lower
level interface gives us a better chance of seeing what's really
in the archive.