rc script: ensure rundir directories created when needed
We were sometimes relying on the pmcd start script being run in
order that PCP_RUN_DIR was created.

In the case of the pmlogger daily script, this assumption could
result in mail being sent to user 'pcp' because the cron script
lacks permissions to create it.  In the case where pmlogger is
not enabled, this results in inappropriate mail.

Resolves Red Hat BZ #1208699.