pmlogcheck: fix basename regression
When I merged the two versions of pmlogcheck I broke the logic
for determining the basename of the PCP archive from the command
line argument.

The problem was exposed by pmmgr which uses pmlogcheck to
validate archives with names like archive-20150415.044829.0
but uses the basename archive-20150415.044829 as a command
line argument to pmlogcheck.  The old code (correctly)
interpreted this as meaning all files matching the pattern
archive-20150415.044829.{index,meta,digit...}, the new code
(incorrectly) thought this was volume 044829 of the archive with
basename archive-20150415 and then failed to find any files
matching the pattern archive-20150415.{index,meta,digit...}
... doh!

This was causing qa/666 to fail, and is fixed in this commit.