pmiostat: usage message cleanup
Noticed pmiostat --help was printing "None" right at the
end, which seemed odd.  Turned out to be a not-quite-right
API call in the usage exception handler.

While there, found the 'x' option not being presented via
the getopt API call (harmless in practice), a double-colon
on "General Options::", and some whitespace/alignment not
quite in the right spot - so, all trivial cleanups here.