pmwebd crasher fix: png-rendered multiple graphs with corrupt archives
It was discovered that if pmwebd (in graphite, render-to-png mode) was
told to draw a mixture of archives, some of which were corrupt enough
to fail pmGetArchiveEnd(), it could crash during the time-series
importance-analysis stage.  We now protect against that particular
case, and improve diagnostics for that and others.  QA included
using two new pcpqa archives, one featuring 100% Genuine Corruption.