libpcp/pmns.c: pmLookupName() fix up
Complete audit of pmLookupName() error handling, to ensure we are
consistent across all of
- local PMNS (with and without a current context)
- remote host context
- local context
- archive context
for all types of metrics, but especially DYNAMIC metrics.

Mostly in libpcp, but also a small pmcd change.

pmLookupName() man page updated to reflect the real error handling
protocol and verified example code added.

As a by-product, discovered that pmNameID() and pmNameAll() were
broken for DYNAMIC metrics, so drive-by fix for these as well.

And finally a bunch of QA consequential cleanup.