pmdamounts: fix stack corruption for large mount options
Miloš found some scenarios where the mounts PMDA could crash
and burn when a machine has mounts with option strings larger
than the allocated space for them.  This commit improves the
bounds checking for all arrays in this PMDA, and also does a
bit of general housecleaning for pmdamount readability.  Test
qa/845 is added to exercise the mount PMDA.

Resolves RHBZ #1213833.