libpcp: avoid pmcpp when loading PMNS if possible
When using pmLoadNameSpace() or pmLoadASCIINameSpace(PM_NS_DEFAULT)
or PM_CONTEXT_LOCAL we don't need pmcpp ... the PMMS contains no
cpp-style #if or #include directives and no cpp macros.

In these cases we use fopen() instead of the fork/exec/pipe
arrangment to pre-process the PMNS file with pmcpp.

The preprocessor is still used with pmLoadASCIINameSpace() and
any PMNS other than PM_NS_DEFAULT, so all of the PMNS updates
via the PMDA Install and Remove scripts still work.