qa: fallout from pmLoadNameSpace() changes
Lots of source code changes here as many QA test applications load
a customized PMNS as part of their testing.  Basically need to
replace pmLoadNameSpace(...) calls with pmLoadASCIINameSpace(..., 1)
calls.

Also a few .out changes as a consequence of not calling any of the
PMNS load namespace routines when we don't need to.