Maintain a list of archives in contexts of type PM_CONTEXT_ARCHIVE.
Within the __pmArchCtl structure, we now maintain a list of active
archives as an array of (__pmLogCtl *). This will allow us to switch
between active archives when the boundaries (begin/end) of the
current archive is reached. The remaining fields of __pmArchCtl
continue to retain their meaning and represent the state of
the current archive. The lower level functions which operate
on the current archive are therefore unchanged at this time.

In order to retain ABI compatibility, the new fields of
pmLogCtl have been added to the end of that struct.

This commit implements the capability to specify a list of archives
to pmNewContext(3), construct the list of archives within the
__pmArchCtl of the context and set the current archive to be the
first one in the list. All single archive semantics are preserved
by these changes.