qa/admin/myconfigure: be smarter and more robust
First, get the Debian configopts directly from debian/rules instead
of replicating the settings here and risking them diverging over
time ... as happened!

Second, for all platforms, since the objective of this script
is to set up the workarea so that Makepkgs and make followed by
"sudo make install" will produce the same results, try and check
that the critical files installed from the package(s) (if found)
match the corresponding files in the workarea after configure
has been run.